Rewrited RangeSlider component
