+++ src/RangeSlider/converters.js
@@ -1,7 +1,7 @@
 export function valueToPosition(value, valuesArray, sliderLength) {
   const index = valuesArray.indexOf(value)
   if (index === -1) {
-    throw new Error(`Invalid value, array does not contain: ${value}`)
+    throw new Error(`Slider: Invalid value, array does not contain: ${value}`)
   }
   const arrLength = valuesArray.length - 1
   return (sliderLength * index) / arrLength
@@ -9,7 +9,7 @@ export function valueToPosition(value, valuesArray, sliderLength) {
 
 export function positionToValue(position, valuesArray, sliderLength) {
   if (position < 0 || sliderLength < position) {
-    throw new Error(`Invalid position: ${position}`)
+    throw new Error(`Slider: Invalid position: ${position}`)
   }
   const arrLength = valuesArray.length - 1
   const index = (arrLength * position) / sliderLength
@@ -18,7 +18,7 @@ export function positionToValue(position, valuesArray, sliderLength) {
 
 export function createArray(start, end, step) {
   if (!step) {
-    throw new Error(`Invalid step: ${step}`)
+    throw new Error(`Slider: Invalid step: ${step}`)
   }
   const length = Math.abs((start - end) / step) + 1
   const direction = start - end > 0 ? -1 : 1
